docs: updating README and adds LICENCE information
Updating the README.md file to provide a better explaination of how the
project works. The page was lacking clarity of what to do with the
project. Introducing some paragraphs to explain how to approach the
repository and start learning TDD.

In addition, the MIT license has been added and linked from the
readme file.

Fixes: #3 #4