feat: introduce code coverage report with codcov.io
Introducing code coverage report for the TDD Traing. This will help
students lookin at their coverage reports.

For a student that wants to have their code analyzed they will be able
to submit a PR to this repo and have Codecov analyze their code and
provide a report.

Note that the setup is not yet implemented properly so it might have
some issue!